web/src/requirements.txt
Flask==2.0.0
google-auth==1.30.0
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.2.0
mysql-connector-python
python-dotenv
google-cloud-storage